refactor: ScaffoldFactory 클래스 -> engine 패키지로 옮김
