+++ src/main/java/me/namuhuchutong/ladder/controller/LadderController.java
@@ -1,7 +1,7 @@
 package me.namuhuchutong.ladder.controller;
 
 import me.namuhuchutong.ladder.domain.*;
-import me.namuhuchutong.ladder.domain.factory.ScaffoldFactory;
+import me.namuhuchutong.ladder.domain.engine.ScaffoldFactory;
 import me.namuhuchutong.ladder.domain.wrapper.Name;
 import me.namuhuchutong.ladder.dto.LadderResult;
 import me.namuhuchutong.ladder.dto.NameAndResult;
