+++ src/main/java/me/namuhuchutong/ladder/domain/Rows.java
@@ -1,6 +1,6 @@
 package me.namuhuchutong.ladder.domain;
 
-import me.namuhuchutong.ladder.domain.factory.ScaffoldFactory;
+import me.namuhuchutong.ladder.domain.engine.ScaffoldFactory;
 
 import java.util.List;
 import java.util.stream.Stream;
