+++ src/main/java/me/namuhuchutong/ladder/domain/engine/ScaffoldFactory.java
@@ -1,4 +1,4 @@
-package me.namuhuchutong.ladder.domain.factory;
+package me.namuhuchutong.ladder.domain.engine;
 
 import me.namuhuchutong.ladder.domain.wrapper.LadderExpression;
